refactor: oauth에 대한 dto 수정으로 인한 리팩토링
